requirements.txt
flask
flask-login
mysql-connector-python